refactor - strategy package 이동
